requirements.txt
flask
sqlalchemy
psycopg2-binary
praw
tweepy
spacy
nltk
vaderSentiment
textblob
transformers
torch
pandas
matplotlib
seaborn
plotly
wordcloud
apscheduler
python-dotenv
requests
protobuf
sentencepiece
langdetect
bs4
newspaper3k
snscrape
flask-cors
lxml_html_clean
httpx
google-generativeai
newsapi-python
google-search-results